BUG: Don't signal FP exceptions (complex absolute)
With the removal of fp-strict, the compiler can potentially use fast-performing
algorithms to compute results. In the case of computing the absolute value of a
complex numpber, the fast inverse square root algorithm is used. Due to the
nature of the algorithm, numpy ends up with a division-by-zero error when
calculating the absolute value of a complex zero.

One could add a condition like `if(in1r == 0.0 && in1i == 0.0)` to check
each element and only clear the float status if this condition is met, but
the compiler would not autovectorize because of it. So instead, numpy is going
to call npy_clear_floatstatus() by default.