BUG: Do not compile pariwise sum with O3 level
As Julian Taylor had written in commit 1e624fa13455e932573b9f421d2865e1fea0cd48,
"pairwise summation has an average error of O(log(n)) instead of O(n) of
the regular summation. It is implemented as summing pairs of small blocks of
regulary summed values in order to archive the same performance as the old
sum."

Because of the way the code was written, compiling it with O3 optimization
and enabling autovectorization can give imprecise results (e.g. with scipy's
test_stats.test_chisquare_masked_arrays test) and the existence of fp-strict
has prevented this from happening even while being compiled with -O3. Now that
it's being removed, we need to make sure the compiler does not make any sort of
optimizations to this function.